+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 0.2.0 - YYYY-MM-DD #
 
-- Plugin "stdin".
+- Disable plugin option.
 
 # 0.1.0 - 2016-02-13 #